Support XMSS hash-based signature aggregation
This PR updates the Signer with an implementation of XMSS to support many-time
signatures using the underlying one-time signature Winternitz scheme.

A Signer uses a HashTree to commit to a sequence of public keys, one
for each slot / epoch.

Several signatures of a common message can be aggregated and verified
together.